+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - Enh #71: Add `HasY2` trait and corresponding tests for managing SVG `y2` attribute (@terabytesoftw)
 - Bug #72: Refactor SVG attribute providers to standardize type validation and improve documentation, and remove redundant excessive tests cases (@terabytesoftw)
 - Bug #73: Update attribute retrieval in tests to use `getAttribute()` method for consistency (@terabytesoftw)
+- Enh #74: Add `Line` class and corresponding tests for SVG `<line>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024